requirements.txt
Flask==2.2.3
matplotlib==3.7.0
numpy==1.26.2
Pillow==10.1.0
python_bcrypt
tensorflow
weasyprint==60.2
